Balanced world 4.
Was very tough due to the high difficulty.

4-4 to 4-7 are particularly difficult.

One some levels, made the sparks much easier to get to balance out the
difficulty. On 4-6 particularly, it gives the player something to do while
the enemies are coming in.

Adrogantia is actually slightly easier than the prior levels if you know
what you're doing. However, the battle is longer, so this seems
appropriate.